sstables: avoid a possible case of unclosed readers
There are some cases of readers being moved into seastar::async, where they are
closed by deferred_close. However, seastar::async may fail with a bad_alloc, and
the deferred_close inside will never run, leaving the reader unclosed and
resulting in an assert violation.

This is not really a stability problem, because if bad_alloc is happening in a
seastar::async, a database failure is imminent anyway. But the assert violation
obscures the bad_alloc and causes confusion.

Fixes #13406