tools: scylla-sstable: don't rely on generating_reader destruction in a thread
The destructors of json_mutation_stream_parser components call .get(),
inducing an assumption that the parser (which is embedded in a
generating_reader) is destroyed in a seastar thread. However, there is no
guarantee about that and following commits break this assumption.

Fix that by performing the asynchronous parts of destruction on end-of-stream,
while we are still inside a future-returning function, rather than in a
destructor.