view_update_generator: Decouple repair and view update processing
When repair writes a sstable to disk, we check if the sstable needs view
update processing. If yes, the sstable will be registered into the view
update processing logic with the _registration_sem semaphore. However,
the semaphore is not very useful because it is too late to add the
back pressure since the sstable is already written.

We have seen multiple cases in the field where the _registration_sem is
stuck so that repair is stuck since repair could not register any more
sstables for view update processing.

This patch decouples repair and view update generator, so repair and
view update processing can work independently.

After this, view update processing is a true background process which
picks sstables in the staging directory and processes.

This makes it easier to reason about which operation is slow or stuck in
the field. It also will make repair significantly faster because repair
now does not wait for view update processing to finish.