db/system_keyspace: add storage for cluster features managed in group 0
The `system.topology` table is extended with a new column
`supported_features` that will hold supported features of a given node.
Although, conceptually, features form a set, the new column will be of
type `text` and will store the features as single string of
comma-separated feature names. This is the convention that is already
used in `system.local`, `system.peers` and also `system.scylla_local`,
so we stick to it for the sake of consistency.