service/raft: pass gms::feature_service and db::system_keyspace
As we are planning to manage features via group 0 state machine, it
needs a reference to the gms::feature_service so that it can enable
features when appropriate. A reference to db::system_keyspace is also
needed so that the feature service can persist the features on disk when
enabling them.