raft topology: include enabled features in snapshot
In order to be sure that the node will be able to properly apply the
contents of the snapshot, it must first make sure that it understands
all of the features that were enabled at the moment when the snapshot
was taken. This information is stored in the system.topology table and
should already be included in topology_mutations of the snapshot, but
that would require parsing them - so, in order to make things simpler, a
new field enabled_features is introduced to the snapshot which
represents the set of currently enabled features as a vector of
sstrings.