doc: mark --load-and-stream as Open Source-only
Fixes https://github.com/scylladb/scylla-enterprise/issues/2807

The --load-and-stream option is included in both Open Source
and Enterprise documentation, while it is only avalilable in
Open Source.
This commit excludes the Load and Stream section from
the Enterprise documentation by marking it as Open Source-only.

This commit must be backported to branch-5.2 and branch-5.1.avalilable