db/hints: ensure_created_and_verified(): also clear hints dir
Hints are supposed to bridge small gaps in availability of nodes.
Keeping hints around across restarts raises problems around staleness of
hints, as it is not possible to if hints are still valid after a
possibly long down-time. Better just wipe them all on startup and let
repair handle cases nodes are down for a prolonged time.