test/cql-pytest: TestScyllaSsstableSchemaLoading: wait for ongoing compactions
The tests in this class rely on no compaction overlapping with their
progress. Despite disabling auto-compaction for the scope of this test
class, we still observed the very first test in the suite failing
sometimes in debug mode. This is most likely a compaction started before
disabling auto-compaction, which (due to the slow debug build) finishes
such that it deletes the compacted sstables, right as one of the tests
runs.
To prevent this, use the recently introduced tasks api to check and wait
for any ongoing compaction before proceeding with the tests.

Fixes: #13553