s3/client: Move last part upload out of finalize_upload()
This change has two reasons. First, is to facilitate moving the
memory_data_sink_buffers from base class, i.e. -- continuation of the
previous patch. Also this fixes a corner case -- if final sink flush
happens right after the previous part was sent for uploading, the
finalization doesn't happen and sink closing aborts the upload even if
it was successful.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>