s3/client: Move memory buffers to upload_sink from base
All the buffers manipulations now happen in the upload_sink class and
the respective member can be removed from base class. The base class
only messes with the buffers in its upload_part() call, but that's
unavoidable, as uploading part implies sending its contents which sits
in buffers.

Now the base class can be re-used for uploading parts with the help of
copy-part API call (next patches)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>