replica/table: wait for materialized view semaphore units
Number of concurrent materialized view updates should be
limited by the view_update_concurrency_semaphore.

For small updates this is achieved using throttling,
the view_update_concurrency_semaphore is used only
to track the memory usage of pending view updates.
The amount by which request responses are throttled
is based on the amount of units taken from the semaphore.

Since the semaphore was used only to keep track
of memory usage, the units were consumed using `seastar::consume`.
`seastar::consume` doesn't wait for the units to be available,
it just decreases the number of units that the semaphore has,
even if they are negative.

Let's change it to `seastar::get_units`, which actually waits for
the units before consuming them.

Having `seastar::consume` causes trouble with queries that generate
a large number of view updates. For small updates throttling is used
to limit the number of concurrent updates, but for big ones the semaphore
should be used.

Imagine a query which deletes a whole partition in the base table.
This is a single query that will generate view updates,
with the number of updates equal to the number of rows in the partition,
potentially in the millions.
In such case throttling doesn't limit the amount of concurent updates at all,
it works on the level of single base table requests.

View updates are generated in a loop, first we generate updates with
at most 100 rows, then consume the units, and then spawn background
tasks which will send them.

Using `seastar::consume` causes unbounded concurrency - the loop
will keep consuming units and generating background view updates
even when the semaphore goes negative.

Changing it to `seastar::get_units` will put a bound on the concurrency,
because now when the semaphore runs out of units no new updates will be spawned
until some of the resources get freed.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>